+++ src/Components/src/Composite/TermSearch/ConfigProvider.fs
@@ -223,6 +223,12 @@ type TermSearchConfigProvider =
         let (activeKeys: TermSearchActiveKeysContext), setActiveKeys =
             React.useLocalStorage (localStorageKey, TermSearchActiveKeysContext.init (defaultActive))
 
+        // From v1.0.7 to v2.0.0 the field was renamed from `aktiveKeys` to `activeKeys`. This migration code ensures that users who have the old field in their local storage will have it removed and replaced with the new field.
+        match activeKeys.activeKeys with
+        | null -> 
+            Browser.Dom.window.localStorage.removeItem(localStorageKey)
+        | _ -> ()
+
         let allKeys =
             React.useMemo (
                 (fun () ->