+++ src/Components/src/Util/Extensions.fs
@@ -20,9 +20,7 @@ type style =
 module Extensions =
     type prop with
         static member inline testid(value: string) : IReactProperty = unbox ("data-testid", value)
-
         static member inline data(key: string, value: string) : IReactProperty = unbox ($"data-{key}", value)
-
         static member inline dataRow(value: int) : IReactProperty = unbox ("data-row", value)
         static member inline dataColumn(value: int) : IReactProperty = unbox ("data-column", value)